- Move class `LdapFluff` to the root of Gem lib-path - Correct `Rakefile` to skip helpers while running unit-tests
