+++ Rakefile
@@ -1,14 +1,12 @@
-require 'rubygems'
+# frozen_string_literal: true
+
+require 'bundler/gem_tasks'
 require 'rake/testtask'
 
-# The default task is run if rake is given no explicit arguments.
-desc 'Default Task'
-task :default => :test
-
-# Test Tasks ---------------------------------------------------------
-
-Rake::TestTask.new('test') do |t|
-  t.libs       = %w[lib test]
-  t.test_files = FileList['test/**/*.rb']
-  t.verbose    = true
+Rake::TestTask.new(:test) do |t|
+  t.libs << 'test'
+  t.libs << 'lib'
+  t.test_files = FileList['test/**/*_test.rb']
 end
+
+task default: :test
